Updated to support newer Ubuntu versions from 24.x

+++ start_wabot.sh
@@ -10,8 +10,10 @@ if [ -e $SCRIPTPATH/.wwebjs_auth/session/Default/Preferences ]; then
 fi
 
 # Remove Preferences-* except for the latest 5 files
-OLDPREFS=$(ls Preferences-*|head --lines=-4)
-[ -z "$OLDPREFS" ] || rm $OLDPREFS
+if [ -e Preferences-* ]; then
+    OLDPREFS=$(ls Preferences-*|head --lines=-4)
+    [ -z "$OLDPREFS" ] || rm $OLDPREFS
+fi
 
 FNLOG=$SCRIPTPATH/wabot.log
 FNSIZE=$(stat -c%s "$FNLOG")
